refactor: remove sspi authentication mechanism
SSPI is the windows GSSAPI implementation, and our two providers
run the exact same handshake. This patch removes the existing
GSSAPI provider, and renames the modern SSPI provider to take its
place.
