+++ lib/core/auth/defaultAuthProviders.js
@@ -4,7 +4,6 @@ const MongoCR = require('./mongocr');
 const X509 = require('./x509');
 const Plain = require('./plain');
 const GSSAPI = require('./gssapi');
-const SSPI = require('./sspi');
 const ScramSHA1 = require('./scram').ScramSHA1;
 const ScramSHA256 = require('./scram').ScramSHA256;
 const MongoDBAWS = require('./mongodb_aws');
@@ -22,7 +21,6 @@ function defaultAuthProviders(bson) {
     x509: new X509(bson),
     plain: new Plain(bson),
     gssapi: new GSSAPI(bson),
-    sspi: new SSPI(bson),
     'scram-sha-1': new ScramSHA1(bson),
     'scram-sha-256': new ScramSHA256(bson)
   };
