+++ lib/core/auth/gssapi.js
@@ -23,7 +23,7 @@ class GSSAPI extends AuthProvider {
       }
     }
 
-    const source = credentials.source;
+    // TODO: Destructure this
     const username = credentials.username;
     const password = credentials.password;
     const mechanismProperties = credentials.mechanismProperties;
@@ -32,199 +32,67 @@ class GSSAPI extends AuthProvider {
       mechanismProperties['gssapiServiceName'] ||
       'mongodb';
 
-    GSSAPIInitialize(
-      this,
-      kerberos.processes.MongoAuthProcess,
-      source,
-      username,
-      password,
-      source,
+    const MongoAuthProcess = kerberos.processes.MongoAuthProcess;
+    const authProcess = new MongoAuthProcess(
+      connection.host,
+      connection.port,
       gssapiServiceName,
-      connection,
-      mechanismProperties,
-      callback
+      mechanismProperties
     );
+
+    authProcess.init(username, password, err => {
+      if (err) return callback(err, false);
+
+      authProcess.transition('', (err, payload) => {
+        if (err) return callback(err, false);
+
+        const command = {
+          saslStart: 1,
+          mechanism: 'GSSAPI',
+          payload,
+          autoAuthorize: 1
+        };
+
+        connection.command('$external.$cmd', command, (err, result) => {
+          if (err) return callback(err, false);
+
+          const doc = result.result;
+          authProcess.transition(doc.payload, (err, payload) => {
+            if (err) return callback(err, false);
+            const command = {
+              saslContinue: 1,
+              conversationId: doc.conversationId,
+              payload
+            };
+
+            connection.command('$external.$cmd', command, (err, result) => {
+              if (err) return callback(err, false);
+
+              const doc = result.result;
+              authProcess.transition(doc.payload, (err, payload) => {
+                if (err) return callback(err, false);
+                const command = {
+                  saslContinue: 1,
+                  conversationId: doc.conversationId,
+                  payload
+                };
+
+                connection.command('$external.$cmd', command, (err, result) => {
+                  if (err) return callback(err, false);
+
+                  const response = result.result;
+                  authProcess.transition(null, err => {
+                    if (err) return callback(err, null);
+                    callback(null, response);
+                  });
+                });
+              });
+            });
+          });
+        });
+      });
+    });
   }
 }
-
-//
-// Initialize step
-var GSSAPIInitialize = function(
-  self,
-  MongoAuthProcess,
-  db,
-  username,
-  password,
-  authdb,
-  gssapiServiceName,
-  connection,
-  options,
-  callback
-) {
-  // Create authenticator
-  var mongo_auth_process = new MongoAuthProcess(
-    connection.host,
-    connection.port,
-    gssapiServiceName,
-    options
-  );
-
-  // Perform initialization
-  mongo_auth_process.init(username, password, function(err) {
-    if (err) return callback(err, false);
-
-    // Perform the first step
-    mongo_auth_process.transition('', function(err, payload) {
-      if (err) return callback(err, false);
-
-      // Call the next db step
-      MongoDBGSSAPIFirstStep(
-        self,
-        mongo_auth_process,
-        payload,
-        db,
-        username,
-        password,
-        authdb,
-        connection,
-        callback
-      );
-    });
-  });
-};
-
-//
-// Perform first step against mongodb
-var MongoDBGSSAPIFirstStep = function(
-  self,
-  mongo_auth_process,
-  payload,
-  db,
-  username,
-  password,
-  authdb,
-  connection,
-  callback
-) {
-  // Build the sasl start command
-  var command = {
-    saslStart: 1,
-    mechanism: 'GSSAPI',
-    payload: payload,
-    autoAuthorize: 1
-  };
-
-  // Write the commmand on the connection
-  connection.command('$external.$cmd', command, (err, result) => {
-    if (err) return callback(err, false);
-
-    // Execute mongodb transition
-    const doc = result.result;
-    mongo_auth_process.transition(doc.payload, function(err, payload) {
-      if (err) return callback(err, false);
-
-      // MongoDB API Second Step
-      MongoDBGSSAPISecondStep(
-        self,
-        mongo_auth_process,
-        payload,
-        doc,
-        db,
-        username,
-        password,
-        authdb,
-        connection,
-        callback
-      );
-    });
-  });
-};
-
-//
-// Perform first step against mongodb
-var MongoDBGSSAPISecondStep = function(
-  self,
-  mongo_auth_process,
-  payload,
-  doc,
-  db,
-  username,
-  password,
-  authdb,
-  connection,
-  callback
-) {
-  // Build Authentication command to send to MongoDB
-  var command = {
-    saslContinue: 1,
-    conversationId: doc.conversationId,
-    payload: payload
-  };
-
-  // Execute the command
-  // Write the commmand on the connection
-  connection.command('$external.$cmd', command, (err, result) => {
-    if (err) return callback(err, false);
-
-    // Call next transition for kerberos
-    const doc = result.result;
-    mongo_auth_process.transition(doc.payload, function(err, payload) {
-      if (err) return callback(err, false);
-
-      // Call the last and third step
-      MongoDBGSSAPIThirdStep(
-        self,
-        mongo_auth_process,
-        payload,
-        doc,
-        db,
-        username,
-        password,
-        authdb,
-        connection,
-        callback
-      );
-    });
-  });
-};
-
-var MongoDBGSSAPIThirdStep = function(
-  self,
-  mongo_auth_process,
-  payload,
-  doc,
-  db,
-  username,
-  password,
-  authdb,
-  connection,
-  callback
-) {
-  // Build final command
-  var command = {
-    saslContinue: 1,
-    conversationId: doc.conversationId,
-    payload: payload
-  };
-
-  // Execute the command
-  connection.command('$external.$cmd', command, (err, result) => {
-    if (err) return callback(err, false);
-
-    const r = result.result;
-    mongo_auth_process.transition(null, function(err) {
-      if (err) return callback(err, null);
-      callback(null, r);
-    });
-  });
-};
-
-/**
- * This is a result from a authentication strategy
- *
- * @callback authResultCallback
- * @param {error} error An error object. Set to null if no error present
- * @param {boolean} result The result of the authentication process
- */
 
 module.exports = GSSAPI;
