+++ lib/operations/connect.js
@@ -34,6 +34,7 @@ const legacyParse = deprecate(
 const AUTH_MECHANISM_INTERNAL_MAP = {
   DEFAULT: 'default',
   PLAIN: 'plain',
+  GSSAPI: 'gssapi',
   'MONGODB-CR': 'mongocr',
   'MONGODB-X509': 'x509',
   'MONGODB-AWS': 'mongodb-aws',
@@ -68,12 +69,12 @@ const monitoringEvents = [
 const VALID_AUTH_MECHANISMS = new Set([
   'DEFAULT',
   'PLAIN',
+  'GSSAPI',
   'MONGODB-CR',
   'MONGODB-X509',
   'MONGODB-AWS',
   'SCRAM-SHA-1',
-  'SCRAM-SHA-256',
-  'GSSAPI'
+  'SCRAM-SHA-256'
 ]);
 
 const validOptionNames = [
@@ -642,16 +643,6 @@ function generateCredentials(client, username, password, options) {
     throw MongoError.create({
       message: `authentication mechanism ${authMechanismRaw} not supported', options.authMechanism`,
       driver: true
-    });
-  }
-
-  if (authMechanism === 'GSSAPI') {
-    return new MongoCredentials({
-      mechanism: process.platform === 'win32' ? 'sspi' : 'gssapi',
-      mechanismProperties: options,
-      source,
-      username,
-      password
     });
   }
